+++ src/app/src/main/java/com/cmput301f21t09/budgetprojectname/UserLoginActivity.java
@@ -11,6 +11,8 @@ import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.cmput301f21t09.budgetprojectname.services.AuthorizationService;
+
 /**
  * Handles sign-in page
  */
