+++ src/app/src/main/java/com/cmput301f21t09/budgetprojectname/services/AuthorizationService.java
@@ -1,4 +1,4 @@
-package com.cmput301f21t09.budgetprojectname;
+package com.cmput301f21t09.budgetprojectname.services;
 
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
